Move Phon memory thunks into weavy
