+++ phon/rust/phon-ir/src/ir.rs
@@ -28,6 +28,10 @@
 
 use phon_schema::bytes::{Reader, skip_pad};
 use phon_schema::{DecodeError, Primitive, SchemaId, SchemaRef};
+pub use weavy::mem::{
+    BorrowThunks, ByteValidator, DefaultOp, DefaultThunk, MapThunks, OpaqueThunks, OptionThunks,
+    PointerThunks, ResultThunks, SeqThunks, SetThunks,
+};
 
 /// A lowered decode program: a straight run of [`Op`]s executed start to finish.
 /// Container bodies (sequence element, map key/value, option payload, enum arm,
@@ -251,28 +255,6 @@ pub enum MemOp {
     CallBlock { schema: SchemaId, offset: usize },
 }
 
-/// A type-erased "write this field's default in place" operation, supplied by the
-/// front door for a reader-only field that can be filled locally
-/// (`#[facet(default)]` or an implicit `Option<T>`/`None` default). The engine
-/// never knows the field type; it calls the thunk, passing back the opaque `ctx` (a
-/// `&'static Shape` for a trait default, or a custom default fn) the front door
-/// understands. Mirrors the `ctx`-carrying thunk style of [`SeqThunks`] and friends
-/// — the spec's bare `fn(slot)` cannot close over the per-field type, so the `ctx`
-/// carries it.
-pub type DefaultThunk = unsafe extern "C" fn(ctx: *const (), slot: *mut u8);
-
-/// A reader-only-default op's payload (boxed in [`MemOp::Default`]). Initializes the
-/// reader field at `base + offset` to its default in place, reading no wire bytes.
-#[derive(Clone, Debug)]
-pub struct DefaultOp {
-    /// Where the reader field lives, relative to the base.
-    pub offset: usize,
-    /// Opaque per-field context the front door binds (passed to `default`).
-    pub ctx: *const (),
-    /// Initialize the uninitialized reader field at `slot` to its default.
-    pub default: DefaultThunk,
-}
-
 /// A pre-built wire skeleton of a writer value, advancing the cursor only — never
 /// reading or writing the reader's memory. Built once at lowering from the writer
 /// schema (see `skip_op`), run by the decode interpreter to consume a writer-only
@@ -402,32 +384,6 @@ pub struct SeqOp {
     pub thunks: SeqThunks,
 }
 
-/// Type-erased operations on an owned sequence handle, supplied by the front door
-/// (`r[descriptors.thunk-binding]`). A `Vec`'s in-memory layout is not something
-/// the engine may assume, so it never pokes the handle directly — it calls these.
-/// `ctx` is an opaque per-type pointer the front door understands (e.g. the
-/// element's list vtable); the engine passes it back untouched.
-///
-/// Decode is engine-owned: the engine allocates and fills the element buffer
-/// itself, then [`from_raw_parts`](Self::from_raw_parts) adopts it into the
-/// handle in one move — no per-element vtable traffic, and in the JIT the alloc
-/// and fill loop are code the engine emits.
-#[derive(Clone, Copy, Debug)]
-pub struct SeqThunks {
-    /// Opaque per-type context, passed to every thunk.
-    pub ctx: *const (),
-    /// Construct the sequence at `list` from a buffer of `len` elements the engine
-    /// allocated with `cap` capacity: `*list = Vec::from_raw_parts(ptr, len, cap)`.
-    /// The buffer must have been allocated with the element type's array layout
-    /// (the engine guarantees this).
-    pub from_raw_parts:
-        unsafe extern "C" fn(ctx: *const (), list: *mut u8, ptr: *mut u8, len: usize, cap: usize),
-    /// The sequence's current element count.
-    pub len: unsafe extern "C" fn(ctx: *const (), list: *const u8) -> usize,
-    /// A pointer to the sequence's contiguous element storage (for reading).
-    pub data: unsafe extern "C" fn(ctx: *const (), list: *const u8) -> *const u8,
-}
-
 /// An owned-set op's payload (boxed in [`MemOp::Set`] to keep `MemOp` small).
 #[derive(Clone, Debug)]
 pub struct SetOp {
@@ -445,41 +401,6 @@ pub struct SetOp {
     /// Type-erased operations on the set handle (front-door bound).
     pub thunks: SetThunks,
 }
-
-/// Type-erased operations on an owned set handle, supplied by the front door
-/// (`r[descriptors.thunk-binding]`). The engine never assumes the set's
-/// in-memory layout: encode iterates borrowed elements, and decode initializes
-/// the set then inserts each decoded element.
-#[derive(Clone, Copy, Debug)]
-pub struct SetThunks {
-    /// Opaque per-type context, passed to every thunk.
-    pub ctx: *const (),
-    /// The set's current element count.
-    pub len: unsafe extern "C" fn(ctx: *const (), set: *const u8) -> usize,
-    /// Initialize the uninitialized set at `set` with room for `cap` entries.
-    pub init_with_capacity: unsafe extern "C" fn(ctx: *const (), set: *mut u8, cap: usize),
-    /// Insert `*value` into the initialized set, moving it out of the scratch
-    /// buffer. Returns `false` when the element was already present.
-    pub insert: unsafe extern "C" fn(ctx: *const (), set: *mut u8, value: *mut u8) -> bool,
-    /// Build a stateful iterator over the initialized set.
-    pub iter_init: unsafe extern "C" fn(ctx: *const (), set: *const u8) -> *mut (),
-    /// Advance the iterator, writing the next borrowed element pointer to
-    /// `value_out` and returning `true`, or returning `false` at the end.
-    pub iter_next:
-        unsafe extern "C" fn(ctx: *const (), iter: *mut (), value_out: *mut *const u8) -> bool,
-    /// Free the iterator built by `iter_init`.
-    pub iter_dealloc: unsafe extern "C" fn(ctx: *const (), iter: *mut ()),
-}
-
-/// Validates a contiguous byte run before it is adopted into an owned handle:
-/// `String` runs check UTF-8 (`r[validate.text]`), `Vec<u8>`/`Vec<scalar>` runs
-/// accept anything. Returns `true` when the bytes are valid for the target type.
-///
-/// One function pointer, called the same way by both engines: the interpreter
-/// invokes it directly, and the JIT reaches it as an *indirect* call (so it needs
-/// no relocation — the reason in-stencil UTF-8 validation routes through here
-/// rather than calling `core::str::from_utf8` inline).
-pub type ByteValidator = unsafe extern "C" fn(ptr: *const u8, len: usize) -> bool;
 
 /// A bulk byte-run op's payload (boxed in [`MemOp::Bytes`]). The wire form is a
 /// `u32` element count then `count * stride` contiguous bytes — one block copy in
@@ -520,31 +441,6 @@ pub struct BorrowOp {
     pub thunks: BorrowThunks,
 }
 
-/// Type-erased operations on a BORROWED contiguous byte run (`&str`/`&[u8]`),
-/// supplied by the front door (`r[descriptors.thunk-binding]`), mirroring
-/// [`SeqThunks`]. The `&str`/`&[T]` fat-pointer layout is unspecified, so the
-/// engine never writes it at a fixed offset — it calls [`set_borrowed`](Self::set_borrowed),
-/// where the type is concrete, to build the fat pointer pointing into the input.
-/// `ctx` is an opaque per-type pointer the engine passes back untouched (it can be
-/// null for the concrete `&str`/`&[u8]` thunks, which need no per-type context).
-#[derive(Clone, Copy, Debug)]
-pub struct BorrowThunks {
-    /// Opaque per-type context, passed to every thunk.
-    pub ctx: *const (),
-    /// Construct the borrowed value at `field`, pointing it at `ptr[..len]` (a run
-    /// INTO the reader's input). For `&str`:
-    /// `core::ptr::write(field.cast::<&str>(), core::str::from_utf8(slice::from_raw_parts(ptr, len))?)`;
-    /// for `&[u8]`: `core::ptr::write(field.cast::<&[u8]>(), slice::from_raw_parts(ptr, len))`.
-    /// Returns `false` on invalid content (e.g. non-UTF-8 for `&str`), which the
-    /// engine maps to a decode error; the field is left uninitialized then.
-    pub set_borrowed:
-        unsafe extern "C" fn(ctx: *const (), field: *mut u8, ptr: *const u8, len: usize) -> bool,
-    /// The borrowed run's element count (its byte length for `&str`/`&[u8]`).
-    pub len: unsafe extern "C" fn(ctx: *const (), field: *const u8) -> usize,
-    /// A pointer to the borrowed run's contiguous bytes (for reading on encode).
-    pub data: unsafe extern "C" fn(ctx: *const (), field: *const u8) -> *const u8,
-}
-
 /// An optional op's payload (boxed in [`MemOp::Option`]). The wire form is a `u8`
 /// presence byte then, only when present, the inner value. The engine never
 /// assumes the in-memory `Option<T>` layout (a repr(Rust) niche or tag); it reads
@@ -564,27 +460,6 @@ pub struct OptionOp {
     pub inner_align: usize,
     /// Type-erased presence operations on the `Option` handle (front-door bound).
     pub thunks: OptionThunks,
-}
-
-/// Type-erased operations on an `Option<T>` handle, supplied by the front door
-/// (`r[descriptors.thunk-binding]`), mirroring [`SeqThunks`]. The engine never
-/// pokes the `Option`'s niche/tag directly — it calls these. `ctx` is an opaque
-/// per-type pointer (the inner type's option vtable) the engine passes back
-/// untouched.
-#[derive(Clone, Copy, Debug)]
-pub struct OptionThunks {
-    /// Opaque per-type context, passed to every thunk.
-    pub ctx: *const (),
-    /// Whether the option at `option` is `Some`.
-    pub is_some: unsafe extern "C" fn(ctx: *const (), option: *const u8) -> bool,
-    /// A pointer to the contained value (valid only when `is_some`).
-    pub get_value: unsafe extern "C" fn(ctx: *const (), option: *const u8) -> *const u8,
-    /// Initialize the uninitialized option at `option` to `Some(*value)`, moving the
-    /// inner value out of `value` (the engine then frees `value`'s storage without
-    /// dropping it).
-    pub init_some: unsafe extern "C" fn(ctx: *const (), option: *mut u8, value: *mut u8),
-    /// Initialize the uninitialized option at `option` to `None`.
-    pub init_none: unsafe extern "C" fn(ctx: *const (), option: *mut u8),
 }
 
 /// A `#[repr(int)]` enum op's payload (boxed in [`MemOp::Enum`]). The wire form is
@@ -649,41 +524,6 @@ pub struct MapOp {
     pub thunks: MapThunks,
 }
 
-/// Type-erased operations on an owned map handle, supplied by the front door
-/// (`r[descriptors.thunk-binding]`), mirroring [`OptionThunks`]. The engine never
-/// pokes the map's in-memory layout directly — it calls these. `ctx` is an opaque
-/// per-type pointer (the map's def) the engine passes back untouched.
-///
-/// Encode is driven by a stateful iterator: `iter_init` builds it, `iter_next`
-/// advances it (yielding borrowed key/value pointers), and `iter_dealloc` frees
-/// it. Decode initializes the map with `init_with_capacity`, then `insert`s each
-/// decoded pair (moving the key and value out of engine scratch).
-#[derive(Clone, Copy, Debug)]
-pub struct MapThunks {
-    /// Opaque per-type context, passed to every thunk.
-    pub ctx: *const (),
-    /// The map's current entry count.
-    pub len: unsafe extern "C" fn(ctx: *const (), map: *const u8) -> usize,
-    /// Initialize the uninitialized map at `map` with room for `cap` entries.
-    pub init_with_capacity: unsafe extern "C" fn(ctx: *const (), map: *mut u8, cap: usize),
-    /// Insert `(*key, *value)` into the initialized map at `map`, moving the key and
-    /// value out of their buffers (the engine then frees both without dropping).
-    pub insert: unsafe extern "C" fn(ctx: *const (), map: *mut u8, key: *mut u8, value: *mut u8),
-    /// Build a stateful iterator over the entries of the initialized map at `map`.
-    pub iter_init: unsafe extern "C" fn(ctx: *const (), map: *const u8) -> *mut (),
-    /// Advance the iterator, writing the next entry's borrowed key and value
-    /// pointers to `key_out`/`value_out` and returning `true`, or returning `false`
-    /// at the end.
-    pub iter_next: unsafe extern "C" fn(
-        ctx: *const (),
-        iter: *mut (),
-        key_out: *mut *const u8,
-        value_out: *mut *const u8,
-    ) -> bool,
-    /// Free the iterator built by `iter_init`.
-    pub iter_dealloc: unsafe extern "C" fn(ctx: *const (), iter: *mut ()),
-}
-
 /// A `Result<T, E>` op's payload (boxed in [`MemOp::Result`]). The wire form is a
 /// `u32` variant index (`ok_wire_index` for `Ok`, `err_wire_index` for `Err`) then
 /// that arm's payload. The engine never assumes the in-memory `Result` layout; it
@@ -716,30 +556,6 @@ pub struct ResultOp {
     pub thunks: ResultThunks,
 }
 
-/// Type-erased operations on a `Result<T, E>` handle, supplied by the front door
-/// (`r[descriptors.thunk-binding]`), mirroring [`OptionThunks`] with two
-/// value-carrying arms. The engine never pokes the `Result`'s niche/tag directly —
-/// it calls these. `ctx` is an opaque per-type pointer (the result's def) the engine
-/// passes back untouched.
-#[derive(Clone, Copy, Debug)]
-pub struct ResultThunks {
-    /// Opaque per-type context, passed to every thunk.
-    pub ctx: *const (),
-    /// Whether the result at `result` is `Ok`.
-    pub is_ok: unsafe extern "C" fn(ctx: *const (), result: *const u8) -> bool,
-    /// A pointer to the contained `Ok` value (valid only when `is_ok`).
-    pub get_ok: unsafe extern "C" fn(ctx: *const (), result: *const u8) -> *const u8,
-    /// A pointer to the contained `Err` value (valid only when not `is_ok`).
-    pub get_err: unsafe extern "C" fn(ctx: *const (), result: *const u8) -> *const u8,
-    /// Initialize the uninitialized result at `result` to `Ok(*value)`, moving the
-    /// inner value out of `value` (the engine then frees `value`'s storage without
-    /// dropping it).
-    pub init_ok: unsafe extern "C" fn(ctx: *const (), result: *mut u8, value: *mut u8),
-    /// Initialize the uninitialized result at `result` to `Err(*value)`, moving the
-    /// inner value out of `value`.
-    pub init_err: unsafe extern "C" fn(ctx: *const (), result: *mut u8, value: *mut u8),
-}
-
 /// An owned-pointer op's payload (boxed in [`MemOp::Pointer`]). The wire form is
 /// the pointee value itself; the owning pointer is local memory detail.
 // r[impl descriptors.thunk-binding]
@@ -756,21 +572,6 @@ pub struct PointerOp {
     pub thunks: PointerThunks,
 }
 
-/// Type-erased operations on an owned pointer handle, supplied by the front door.
-/// The engine never assumes the pointer layout or allocation strategy: it borrows
-/// the pointee for encode and constructs the owner from a decoded pointee on
-/// decode.
-// r[impl descriptors.thunk-binding]
-#[derive(Clone, Copy, Debug)]
-pub struct PointerThunks {
-    /// Opaque per-type context, passed to every thunk.
-    pub ctx: *const (),
-    /// Borrow the initialized pointer's pointee.
-    pub borrow: unsafe extern "C" fn(ctx: *const (), pointer: *const u8) -> *const u8,
-    /// Initialize `pointer` from `*value`, moving the pointee out of engine scratch.
-    pub init: unsafe extern "C" fn(ctx: *const (), pointer: *mut u8, value: *mut u8),
-}
-
 /// An opaque-field op's payload (boxed in [`MemOp::Opaque`]). The wire form is a
 /// `u32` byte length then that many inner bytes — IDENTICAL to a `Primitive::Bytes`
 /// run, so a peer that does not know the inner type reads it as opaque bytes. The
@@ -782,32 +583,6 @@ pub struct OpaqueOp {
     pub field_offset: usize,
     /// Type-erased encode/decode of the inner value (front-door bound).
     pub thunks: OpaqueThunks,
-}
-
-/// Type-erased operations on an opaque field (`#[facet(opaque = ...)]`), supplied
-/// by the front door (`r[descriptors.thunk-binding]`), mirroring [`SeqThunks`]. The
-/// engine never knows the inner type — it frames the field as a length-prefixed
-/// blob and delegates the inner bytes to these thunks. `ctx` is an opaque per-field
-/// pointer (the field's opaque-adapter definition) the engine passes back untouched.
-///
-/// Encode reserves a `u32` length, calls [`encode`](Self::encode) to APPEND the
-/// inner value's bytes to `out`, then backpatches the length. Decode reads the
-/// length, borrows the inner span from the reader's input, and calls
-/// [`decode`](Self::decode) to build the value at `slot` — the decoded value may
-/// borrow that span (the standard zero-copy contract).
-#[derive(Clone, Copy, Debug)]
-pub struct OpaqueThunks {
-    /// Opaque per-field context, passed to every thunk.
-    pub ctx: *const (),
-    /// Append the inner value's encoded bytes to `out`. `field` points at the
-    /// opaque field in memory; the engine has already reserved (and will backpatch)
-    /// the `u32` length prefix, so this writes ONLY the inner bytes.
-    pub encode: unsafe extern "C" fn(ctx: *const (), field: *const u8, out: *mut Vec<u8>),
-    /// Build the opaque value at `slot` from the inner span `bytes[..len]` (borrowed
-    /// from the reader's input). Returns `false` if the adapter rejects the input,
-    /// which the engine maps to a decode error (the field is left uninitialized then).
-    pub decode:
-        unsafe extern "C" fn(ctx: *const (), bytes: *const u8, len: usize, slot: *mut u8) -> bool,
 }
 
 /// Coalesce adjacent scalar copies that are contiguous in *both* the wire and
